+++ src/internal/interpreter/map-interpreter.errors.test.ts
@@ -83,9 +83,9 @@ describe('MapInterpreter errors', () => {
         },
         { fetchInstance }
       );
-      const result = await interpreter.perform(({
+      const result = await interpreter.perform({
         kind: 'Invalid',
-      } as unknown) as MapDocumentNode);
+      } as unknown as MapDocumentNode);
       expect(result.isErr() && result.error instanceof UnexpectedError).toEqual(
         true
       );